Simplify the implementations
