--- src/nn/compilation_options.ts
@@ -1,9 +1,0 @@
-import {PowerPreference} from './power_preference';
-
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#dictdef-compilationoptions)
- */
-export interface CompilationOptions {
-  /** */
-  powerPreference?: PowerPreference;
-}
