--- src/nn/input_operand.ts
@@ -1,17 +1,0 @@
-import {ModelBuilder} from './model_builder_impl';
-import {OperandDescriptor} from './operand_descriptor';
-import {Operand} from './operand_impl';
-import {assert, validateOperandDescriptor} from './utils';
-
-export class InputOperand extends Operand {
-  readonly name: string;
-  readonly desc: OperandDescriptor;
-
-  constructor(name: string, desc: OperandDescriptor, builder: ModelBuilder) {
-    super(builder);
-    assert(typeof name === 'string', 'The name parameter is invalid');
-    this.name = name;
-    validateOperandDescriptor(desc);
-    this.desc = desc;
-  }
-}
