--- src/nn/named_inputs.ts
@@ -1,6 +1,0 @@
-import {Input} from './input';
-
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#typedefdef-namedinputs)
- */
-export type NamedInputs = Record<string, Input>;
