--- src/nn/named_operands.ts
@@ -1,6 +1,0 @@
-import {Operand} from './operand';
-
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#typedefdef-namedoperands)
- */
-export type NamedOperands = Record<string, Operand>;
