--- src/nn/named_outputs.ts
@@ -1,6 +1,0 @@
-import {Output} from './output';
-
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#typedefdef-namedoutputs)
- */
-export type NamedOutputs = Record<string, Output>;
