+++ src/nn/neural_network_context.ts
@@ -1,18 +1,10 @@
 import {ModelBuilder} from './model_builder';
-import {NeuralNetworkContext as NeuralNetworkContextImpl} from './neural_network_context_impl';
-
 
 /**
  * [spec](https://webmachinelearning.github.io/webnn/#api-neuralnetworkcontext)
  */
-export interface NeuralNetworkContext {
-  /** */
-  createModelBuilder(): ModelBuilder;
+export class NeuralNetworkContext {
+  createModelBuilder(): ModelBuilder {
+    return new ModelBuilder();
+  }
 }
-
-interface NeuralNetworkContextConstructor {
-  new(): NeuralNetworkContext;
-}
-// eslint-disable-next-line no-redeclare
-export const NeuralNetworkContext: NeuralNetworkContextConstructor =
-    NeuralNetworkContextImpl;
