--- src/nn/operand_layout.ts
@@ -1,7 +1,0 @@
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#enumdef-operandlayout)
- */
-export enum OperandLayout {
-  'nchw' = 'nchw',
-  'nhwc' = 'nhwc'
-}
