--- src/nn/operand_type.ts
@@ -1,11 +1,0 @@
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#enumdef-operandtype)
- */
-export enum OperandType {
-  'float32' = 'float32',
-  'float16' = 'float16',
-  'int32' = 'int32',
-  'uint32' = 'uint32',
-  'int8' = 'int8',
-  'uint8' = 'uint8'
-}
