--- src/nn/output.ts
@@ -1,11 +1,0 @@
-import {ArrayBufferView} from './types';
-
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#dictdef-output)
- */
-export interface Output {
-  /** */
-  readonly buffer?: ArrayBufferView;
-  /** */
-  readonly dimensions?: number[];
-}
