--- src/nn/power_preference.ts
@@ -1,8 +1,0 @@
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#enumdef-powerpreference)
- */
-export enum PowerPreference {
-  'default' = 'default',
-  'high-performance' = 'high-performance',
-  'low-power' = 'low-power'
-}
