--- src/nn/slice_options.ts
@@ -1,6 +1,0 @@
-/**
- * [spec](https://webmachinelearning.github.io/webnn/#dictdef-sliceoptions)
- */
-export interface SliceOptions {
-  axes?: number[];
-}
